notify-service/requirements.txt
pytz
python-dateutil
psycopg2-binary
dotenv
pytest
python-dotenv